requirements.txt
streamlit>=1.32.0
Pillow>=10.0.0
requests>=2.31.0
python-dotenv>=1.0.1
streamlit-authenticator>=0.2.2
watchdog>=4.0.0
pandas>=2.2.1
opencv-python>=4.9.0.80
